requirements.txt
#requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
gtts==2.4.0
feedparser==6.0.10
pydantic==2.5.0
plotly==5.17.0
pandas==2.1.3